+++ src/CheckoutCart.js
@@ -4,19 +4,21 @@ import { Link } from 'react-router-dom';
 import Checkout from './Checkout';
 
 
-class Cart extends React.Component {
+class CheckoutCart extends React.Component {
 
 
   render() {
     
   const cartItems = this.props.cartItems.map((item, i) =>
-    <CartItem key={item.code} item={item} onRemoveFromCart={this.props.onRemoveFromCart} />
+  <tr>
+      <td>{item.name}</td>
+      <td>{item.quantity}</td>
+      <td><span>$</span>{item.totalPrice.toFixed(2)}</td>
+     
+    </tr>
   );
-
   const totalPrice = this.props.totalPrice.toFixed(2);
-  const checkout = this.props.cartItems.map((item, i) =>
-    <CartItem key={item.code} item={item} onRemoveFromCart={this.props.onRemoveFromCart} />
-  );
+
   return (
     <div className="cart">
       <table>
@@ -25,7 +27,7 @@ class Cart extends React.Component {
             <td>Name</td>
             <td>Quantity</td>
             <td>Price</td>
-            <td>Action</td>
+           
           </tr>
         </thead>
         <tbody>
@@ -35,11 +37,9 @@ class Cart extends React.Component {
       <h2 className="cart-total-price">
         {/* Dollar sign below is just for the unit, not string interpolation. */}
         Total: <span>${totalPrice === (-0.00) ? 0.00 : totalPrice}</span>
-        <Link to="/Checkout">
-     <button id="button"type="button">
-          Click Me!
-     </button>
- </Link>
+    
+
+
       </h2>
     </div>
   );
@@ -47,4 +47,4 @@ class Cart extends React.Component {
 }
 }
 
-export default Cart;
+export default CheckoutCart;
