+++ src/index.js
@@ -1,10 +1,13 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
+import { BrowserRouter } from 'react-router-dom';
 
 import App from './App.js';
 import data from './data.js';
 
 ReactDOM.render(
-  <App data={data} />,
-  document.getElementById('root')
+  <BrowserRouter>
+  <App data={data} />
+  </BrowserRouter>
+  , document.getElementById('root')
 );
